+++ src/vegapunk/simulators/fetorch/element/type/hexa20.py
@@ -119,7 +119,7 @@ class FEHexa20(ElementType):
                           (1.0, 0.0, 1.0),
                           (0.0, 1.0, 1.0),
                           (-1.0, 0.0, 1.0)],
-                         dtype=torch.float, device=self._device)
+                         device=self._device)
         # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         # Store nodes local coordinates
         self._nodes_local_coords = nodes_local_coords
